Fixing various warnings in tests/testglib.c
tests/testglib.c: In function ‘glist_test’:
tests/testglib.c:90:30: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
       if (*((gint*) t->data) != (9 - i))
                              ^~
tests/testglib.c:95:54: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
     if (g_list_position (list, g_list_nth (list, i)) != i)
                                                      ^~
tests/testglib.c:112:30: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
       if (*((gint*) t->data) != i)
                              ^~
tests/testglib.c:130:30: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
       if (*((gint*) t->data) != (9 - i))
                              ^~
tests/testglib.c:150:30: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
       if (*((gint*) t->data) != (9 - i))
                              ^~
tests/testglib.c: In function ‘gslist_test’:
tests/testglib.c:170:31: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
       if (*((gint*) st->data) != (9 - i))
                               ^~
tests/testglib.c:188:31: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
       if (*((gint*) st->data) != i)
                               ^~
tests/testglib.c:206:31: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
       if (*((gint*) st->data) != (9 - i))
                               ^~
tests/testglib.c:226:31: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
       if (*((gint*) st->data) != (9 - i))
                               ^~
In file included from glib/glibconfig.h:9,
                 from glib/gtypes.h:32,
                 from glib/galloca.h:32,
                 from glib/glib.h:30,
                 from tests/testglib.c:31:
tests/testglib.c: In function ‘gnode_test’:
tests/testglib.c:302:34: error: comparison of integer expressions of different signedness: ‘char’ and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
       g_assert (P2C (node->data) == ('C' + i));
                                  ^~
glib/gmacros.h:455:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
tests/testglib.c:302:7: note: in expansion of macro ‘g_assert’
       g_assert (P2C (node->data) == ('C' + i));
       ^~~~~~~~
tests/testglib.c:306:76: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
     g_assert (g_node_child_position (node_G, g_node_nth_child (node_G, i)) == i);
                                                                            ^~
glib/gmacros.h:455:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
tests/testglib.c:306:5: note: in expansion of macro ‘g_assert’
     g_assert (g_node_child_position (node_G, g_node_nth_child (node_G, i)) == i);
     ^~~~~~~~
tests/testglib.c: In function ‘test_arrays’:
tests/testglib.c:1123:41: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
     if (g_array_index (garray, gint, i) != i)
                                         ^~
tests/testglib.c:1131:41: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
     if (g_array_index (garray, gint, i) != (100 - i - 1))
                                         ^~

+++ tests/testglib.c
@@ -78,7 +78,7 @@ static void
 glist_test (void)
 {
   GList *list = NULL;
-  guint i;
+  gint i;
 
   for (i = 0; i < 10; i++)
     list = g_list_append (list, &test_nums[i]);
@@ -158,7 +158,7 @@ static void
 gslist_test (void)
 {
   GSList *slist = NULL;
-  guint i;
+  gint i;
 
   for (i = 0; i < 10; i++)
     slist = g_slist_append (slist, &test_nums[i]);
@@ -299,11 +299,11 @@ gnode_test (void)
   for (i = 0; i < g_node_n_children (node_B); i++)
     {
       node = g_node_nth_child (node_B, i);
-      g_assert (P2C (node->data) == ('C' + i));
+      g_assert (P2C (node->data) == ('C' + (gint) i));
     }
   
   for (i = 0; i < g_node_n_children (node_G); i++)
-    g_assert (g_node_child_position (node_G, g_node_nth_child (node_G, i)) == i);
+    g_assert (g_node_child_position (node_G, g_node_nth_child (node_G, i)) == (gint) i);
 
   /* we have built:                    A
    *                                 /   \
@@ -1094,7 +1094,7 @@ test_arrays (void)
   GByteArray *gbarray;
   GPtrArray *gparray;
   GArray *garray;
-  guint i;
+  gint i;
 
   gparray = g_ptr_array_new ();
   for (i = 0; i < 10000; i++)
